requirements.txt
fastapi>=0.111
pydantic>=2.7
PyYAML>=6.0
uvicorn>=0.30
pytest>=8.0
